move generic flag types to flagutil
