+++ cmd/earthly/subcmd/common.go
@@ -2,56 +2,10 @@ package subcmd
 
 import (
 	"context"
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/dustin/go-humanize"
 	"github.com/earthly/earthly/cloud"
 	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
 )
-
-type byteSizeValue uint64
-
-func (b *byteSizeValue) Set(s string) error {
-	v, err := humanize.ParseBytes(s)
-	if err != nil {
-		return err
-	}
-	*b = byteSizeValue(v)
-	return nil
-}
-
-func (b *byteSizeValue) String() string { return humanize.Bytes(uint64(*b)) }
-
-// duration implements cli.GenericFlag methods to support time.Duration with days, e.g. 1d
-type duration time.Duration
-
-func (d *duration) String() string {
-	return (time.Duration(*d)).String()
-}
-
-func (d *duration) Set(value string) error {
-	if value == "" {
-		return nil
-	}
-	daysToHours := false
-	if strings.HasSuffix(value, "d") {
-		value = fmt.Sprintf("%s%s", strings.TrimSuffix(value, "d"), "h")
-		daysToHours = true
-	}
-	dur, err := time.ParseDuration(value)
-	if err != nil {
-		return errors.New("parse error")
-	}
-
-	if daysToHours {
-		dur *= 24
-	}
-	*d = duration(dur)
-	return nil
-}
 
 // projectOrgName returns the specified org or retrieves the default org from the API.
 func projectOrgName(cli CLI, ctx context.Context, cloudClient *cloud.Client) (string, error) {
